Attempt to force MySQL to use index on TranslationGroupID instead of SiteTree.Sort
For the translation lookup query in `getTranslations`, we noticed that MySQL
would sometimes use a query plan that would `ORDER BY` the `SiteTree.Sort`
field and then limit by `TranslationGroupID`. This query plan made this query,
that should run in <10ms, take >100sec depending on the size of SiteTree. This
commit adds a hack to force MySQL to use the `TranslationGroupID` index rather
than the `SiteTree.Sort` index.